IClient & Commands
I updated IClient and the Client Facade and made some new commands. These new commands can all be found in the InGameCommands package. One thing we still need to implement is a way to mark a claimed route with markers of the player's color. The changes to IServer and the implementation of chat will be next.